requirements.txt
Flask~=3.1.0
Flask-Babel~=4.0.0
Flask-WTF~=1.2.2
Flask-SQLAlchemy~=3.1.1
WTForms~=3.2.1
python-telegram-bot~=21.10
